# backend/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Facturo API",
        "description": "API REST : authentification (Sanctum), clients, devis, factures, synthèse tableau de bord.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "API locale — aligner avec APP_URL et L5_SWAGGER_CONST_HOST dans .env"
        }
    ],
    "paths": {
...
        "/api/quotes": {
            "get": {
                "tags": [
                    "Devis"
                ],
                "operationId": "041a779c7ea82128b78050569aaf33e0",
                "responses": {
                    "200": {
                        "description": "Liste paginée (items inclus)"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Devis"
                ],
                "summary": "Créer un devis (lignes optionnelles : recalcul des totaux)",
                "operationId": "6fe559a8e09a68c619d27371d3c6bed3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "number"
                                ],
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "nullable": true,
                                        "enum": [
                                            "draft",
                                            "sent",
                                            "accepted",
                                            "rejected"
                                        ]
                                    },
                                    "issue_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "valid_until": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "example": "XOF",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "description": {
                                                    "type": "string"
                                                },
                                                "quantity": {
                                                    "type": "number"
                                                },
                                                "unit_price": {
                                                    "type": "number"
                                                },
                                                "tax_rate": {
                                                    "type": "number",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Devis créé"
                    },
                    "422": {
                        "description": "Validation"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/quotes/{id}": {
            "get": {
                "tags": [
                    "Devis"
                ],
                "operationId": "148d8128c3eb0f4d588b08d654df96f0",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Détail"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Devis"
                ],
                "summary": "Mettre à jour (si `items` est envoyé, lignes remplacées et totaux recalculés)",
                "operationId": "3e7cd5844132930b61a4d6379447a5e3",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "client_id": {
                                        "type": "integer",
                                        "nullable": true
                                    },
                                    "number": {
                                        "type": "string"
                                    },
                                    "status": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "issue_date": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "valid_until": {
                                        "type": "string",
                                        "format": "date",
                                        "nullable": true
                                    },
                                    "currency": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "notes": {
                                        "type": "string",
                                        "nullable": true
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "type": "object"
                                        },
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Mis à jour"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Devis"
                ],
                "operationId": "aab4d804df512622defaf39dad61ab5c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Supprimé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        }
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token Sanctum : header Authorization: Bearer {token}",
                "bearerFormat": "Token",
                "scheme": "bearer"
            }
        }
    },
...
}